refactor: マージする手続きも、Domain側に持っていく。
整理できてないから、ただただコードが移動しただけなのはダメだと思うが。
